Implemented Phase for ScrollTimelines
Implemented Phase for ScrollTimelines. Also updated Phase to match
current time model for internal and external behavior which will allow
phase to be accessed as an enum in animation.cc where it will be used
in different calculations.

Added unit test as well as a WPT test for ScrollTimeline Phases.

Bug: 1046833
Change-Id: Ib86f68b34ef91f4e70fdefd55bdb2629dc415a72