Adds validity check to imperative slot API. Throws exception when failed.
Prior to this CL, calling imperative slot API does not check the
validity of the slot element or its assigned nodes.

This CL validates the slot element and its assigned nodes when calling
imperative slot API. The checks are: shadow root is V1, shadow root is
in manual slot assignment mode, and assign nodes are children of slot's
host element.

Bug:869308
Change-Id: I4b0a1a1a3af6bf159e22af5eafe32ed44172b021